fix: It is executed twice when the log parameter is a function
The current format checking method will cause the parameters to be used by two functions, so if an expression is passed in, it will be executed twice.

This is a bug, and then need to remove all these functions of C and use C++ template to match the printing parameters.